+++ app/src/main/java/lavender/client/android/data/updates/UpdateUtils.kt
@@ -54,19 +54,8 @@ object UpdateUtils {
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
 
-        // Install button intent
         val installButtonPendingIntent = PendingIntent.getActivity(
             context, 1005, installIntent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-
-        // Intent to show changelog
-        val whatsNewIntent = Intent(context, lavender.client.android.ui.chatlist.ChatListActivity::class.java).apply {
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
-            putExtra("extra_show_whats_new", true)
-        }
-        val whatsNewPendingIntent = PendingIntent.getActivity(
-            context, 1004, whatsNewIntent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
 
@@ -77,7 +66,6 @@ object UpdateUtils {
             .setPriority(NotificationCompat.PRIORITY_HIGH)
             .setContentIntent(pendingIntent)
             .addAction(R.drawable.ic_install_update, context.getString(R.string.install_now), installButtonPendingIntent)
-            .addAction(R.drawable.ic_star, context.getString(R.string.whats_new), whatsNewPendingIntent)
             .setAutoCancel(true)
             .build()